octeontx2-pf: Fix coverity issues
commit eaf507a5e93c111f3a6a33b00d16fa3c9001cd5e from
git@git.assembla.com:cavium/WindRiver.linux.git

 Fix unintended sign extension issues.

Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/94549
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>

+++ drivers/net/ethernet/marvell/octeontx2/nic/qos.c
@@ -131,7 +131,7 @@ static void __otx2_qos_txschq_cfg(struct otx2_nic *pfvf,
 
 		/* configure parent txschq */
 		cfg->reg[num_regs] = NIX_AF_MDQX_PARENT(node->schq);
-		cfg->regval[num_regs] = node->parent->schq << 16;
+		cfg->regval[num_regs] = (u64)node->parent->schq << 16;
 		num_regs++;
 
 		/* configure prio/quantum */
@@ -205,7 +205,7 @@ static void __otx2_qos_txschq_cfg(struct otx2_nic *pfvf,
 				  node->quantum : pfvf->tx_max_pktlen;
 			rr_weight = otx2_qos_quantum_to_dwrr_weight(pfvf,
 								    quantum);
-			cfg->regval[num_regs] = node->parent->child_dwrr_prio << 24 |
+			cfg->regval[num_regs] = (u64)node->parent->child_dwrr_prio << 24 |
 						rr_weight;
 		}
 		num_regs++;
@@ -312,7 +312,7 @@ static void __otx2_qos_txschq_cfg(struct otx2_nic *pfvf,
 		/* check if node is root */
 		if (node->qid == OTX2_QOS_QID_INNER && !node->parent) {
 			cfg->reg[num_regs] = NIX_AF_TL2X_SCHEDULE(node->schq);
-			cfg->regval[num_regs] =  hw->txschq_aggr_lvl_rr_prio << 24 |
+			cfg->regval[num_regs] =  (u64)hw->txschq_aggr_lvl_rr_prio << 24 |
 						 mtu_to_dwrr_weight(pfvf,
 								    pfvf->tx_max_pktlen);
 			num_regs++;
@@ -329,7 +329,7 @@ static void __otx2_qos_txschq_cfg(struct otx2_nic *pfvf,
 				  node->quantum : pfvf->tx_max_pktlen;
 			rr_weight = otx2_qos_quantum_to_dwrr_weight(pfvf,
 								    quantum);
-			cfg->regval[num_regs] = node->parent->child_dwrr_prio << 24 |
+			cfg->regval[num_regs] = (u64)node->parent->child_dwrr_prio << 24 |
 						rr_weight;
 		}
 		num_regs++;
